requirements.txt
websockets>=12.0
aiohttp>=3.9.0
python-telegram-bot>=20.0
